ai_vision/requirements.txt
ultralytics
opencv-python==4.10.0.84
python-dotenv
requests
inference-sdk